+++ .env.example
@@ -1,3 +1,3 @@
-PORT=
-SALT=
-DB_HOST=
+PORT=4000
+SALT="super secret"
+DB_HOST="127.0.0.1"
