--- src/config/index.ts
@@ -1,2 +1,0 @@
-export {RestConfig} from './rest.js';
-export {configRestSchema,RestAppConfig} from './rest-schema.js';
