+++ src/config/rest/rest.schema.ts
@@ -1,21 +1,13 @@
-import convict from '#lib/convict.js';
+import convict from 'convict';
 
-import { Config } from './interface.js';
-
-export type RestSchema = {
-  PORT: number;
-  SALT: string;
-  DB_HOST: string;
-}
-
-export type RestAppConfig = Config<RestSchema>
+import { RestSchema } from './rest.types.js';
 
 export const configRestSchema = convict<RestSchema>({
   PORT: {
     doc: 'Port for incoming connections',
     format: 'port',
     env: 'PORT',
-    default: 4000
+    default: null
   },
   SALT: {
     doc: 'Salt for password hash',
@@ -27,6 +19,6 @@ export const configRestSchema = convict<RestSchema>({
     doc: 'IP address of the database server (MongoDB)',
     format: 'ipaddress',
     env: 'DB_HOST',
-    default: '127.0.0.1'
+    default: null
   },
 });
