+++ src/config/rest/rest.types.ts
@@ -1,0 +1,9 @@
+import { Config } from '#config/config.interface.js';
+
+export type RestSchema = {
+  PORT: number;
+  SALT: string;
+  DB_HOST: string;
+}
+
+export type RestAppConfig = Config<RestSchema>
