+++ src/lib/file-reader/file-reader.interface.ts
