+++ src/lib/file-reader/index.ts
@@ -1,1 +1,1 @@
-export { TSVFileReader } from './tsv-reader.js';
+export { TSVFileReader } from './tsv-file-reader.js';
