+++ src/lib/file-reader/tsv-file-reader.ts
@@ -1,9 +1,9 @@
 import EventEmitter from 'node:events';
 import { createReadStream} from 'node:fs';
 
-import { FileReader } from './reader.js';
+import { FileReader } from './file-reader.interface.js';
 
-const CHUNK_SIZE = 16384; // 16KB
+const CHUNK_SIZE = 16384;
 
 export class TSVFileReader extends EventEmitter implements FileReader {
   constructor(
