+++ src/lib/file-writer/tsv-file-writer.ts
@@ -1,6 +1,6 @@
 import { createWriteStream, WriteStream } from 'node:fs';
 
-import { FileWriter } from './writer.js';
+import { FileWriter } from './file-writer.interface.js';
 
 export class TSVFileWriter implements FileWriter {
   private stream: WriteStream;
