--- src/lib/logger.ts
@@ -1,58 +1,0 @@
-import {dirname, resolve} from 'node:path';
-import { fileURLToPath } from 'node:url';
-
-import { injectable } from 'inversify';
-import { Logger as PinoInstance, pino, transport } from 'pino';
-
-export interface Logger {
-  info(message: string, ...args: unknown[]): void;
-  warn(message: string, ...args: unknown[]): void;
-  error(message: string, error: Error, ...args: unknown[]): void;
-  debug(message: string, ...args: unknown[]): void;
-}
-
-const LOG_FILE_PATH = 'logs/rest.log';
-
-const __dirname = dirname(fileURLToPath(import.meta.url));
-
-@injectable()
-export class PinoLogger implements Logger {
-  private readonly logger: PinoInstance;
-
-  constructor() {
-    const destination = resolve(__dirname, '../../', LOG_FILE_PATH);
-
-    const multiTransport = transport({
-      targets: [
-        {
-          target: 'pino/file',
-          options: { destination },
-          level: 'debug'
-        },
-        {
-          target: 'pino/file',
-          level: 'info',
-        }
-      ],
-    });
-
-    this.logger = pino({}, multiTransport);
-    this.logger.info('Logger created…');
-  }
-
-  public debug(message: string, ...args: unknown[]): void {
-    this.logger.debug(message, ...args);
-  }
-
-  public error(message: string, error: Error, ...args: unknown[]): void {
-    this.logger.error(error, message, ...args);
-  }
-
-  public info(message: string, ...args: unknown[]): void {
-    this.logger.info(message, ...args);
-  }
-
-  public warn(message: string, ...args: unknown[]): void {
-    this.logger.warn(message, ...args);
-  }
-}
