+++ src/lib/logger/index.ts
@@ -1,0 +1,2 @@
+export {Logger} from './logger.interface.js';
+export {PinoLogger} from './pino.logger.js';
