+++ src/lib/logger/logger.interface.ts
@@ -1,0 +1,6 @@
+export interface Logger {
+  info(message: string, ...args: unknown[]): void;
+  warn(message: string, ...args: unknown[]): void;
+  error(message: string, error: Error, ...args: unknown[]): void;
+  debug(message: string, ...args: unknown[]): void;
+}
