+++ src/lib/logger/pino.logger.ts
@@ -1,0 +1,30 @@
+import { injectable } from 'inversify';
+import { Logger as PinoInstance, pino, } from 'pino';
+
+import { Logger } from './logger.interface.js';
+
+@injectable()
+export class PinoLogger implements Logger {
+  private readonly logger: PinoInstance;
+
+  constructor() {
+    this.logger = pino();
+    this.logger.info('Logger created…');
+  }
+
+  public debug(message: string, ...args: unknown[]): void {
+    this.logger.debug(message, ...args);
+  }
+
+  public error(message: string, error: Error, ...args: unknown[]): void {
+    this.logger.error(error, message, ...args);
+  }
+
+  public info(message: string, ...args: unknown[]): void {
+    this.logger.info(message, ...args);
+  }
+
+  public warn(message: string, ...args: unknown[]): void {
+    this.logger.warn(message, ...args);
+  }
+}
