+++ src/lib/offer-generator/offer-generator.interface.ts
