+++ src/lib/offer-generator/tsv-offer-generator.enum.ts
