+++ src/rest/index.ts
@@ -1,0 +1,1 @@
+export {RestApp} from './rest.app.js';
