+++ src/types/comment.types.ts
@@ -1,4 +1,4 @@
-import { User } from './user.js';
+import { User } from './user.types.js';
 
 export type Comment = {
   text: string;
