+++ src/types/component.enum.ts
