+++ src/types/offer-type.enum.ts
@@ -1,0 +1,6 @@
+export enum OfferType {
+  APARTMENT = 'apartment',
+  HOUSE = 'house',
+  ROOM = 'room',
+  HOTEL = 'hotel'
+}
