+++ src/types/user-account.enum.ts
@@ -1,0 +1,4 @@
+export enum UserAccountType {
+  COMMON = 'обычный',
+  PRO = 'pro',
+}
